financial_data_code_for_refernce/requirements.txt
# Core Dependencies
anthropic==0.34.2
python-dotenv==1.0.0
pymupdf==1.24.1

# Optional (for frontend later)
streamlit==1.28.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
